This is a STRONGLY recommended update.
The issue, found by Matthieu Huguet, comes from Token/Consumer.generate_random_codes  functions which test the key AND secret combination and not key OR secret. This is not a security issue because stores.DataStore.lookup_consumer/lookup_token tries to retrieve the token with an objects.get so it will raise a model.MultipleObjectsReturned error in case there are two similar keys. But you must be careful if you have customized the store.

Please contact me (http://larlet.com) if you'd like to be added to the django-oauth-security mailing-list for future security announcements.